requirements.txt
Django
djangorestframework
psycopg2-binary
drf-spectacular
pandas
openpyxl
requests
python-dotenv
whitenoise